Corregí datos de precios el la db y models. Agruegé el precio deempanadas en SeleccionProductos.ejs

+++ database/models/PrecioEmpanada.js
@@ -10,7 +10,7 @@ module.exports = (sequelize, dataTypes) => {
       type: dataTypes.INTEGER
     },
     precio: {
-      type: dataTypes.DECIMAL(10, 2)
+      type: dataTypes.INTEGER(10)
     }
   }
   const config = {
